% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerScale.R
\name{centerScale}
\alias{centerScale}
\title{centerScale uses the Mean Centering Normalisation method to scale the dataset}
\usage{
centerScale(xMatrix)
}
\arguments{
\item{xMatrix}{matrix of data to be scaled}
}
\value{
returns the variable scalex, which is the scaled matrix
}
\description{
centerScale uses the Mean Centering Normalisation method to scale the dataset
}
\details{
This function uses the equation x' = x - mean(x) to mean center scale all data within the matrix.
}
