% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doKNN.R
\name{doKNN}
\alias{doKNN}
\title{doKNN Applies the K-Nearest Neighbours Regression model to the training data,
and calculates the Root Mean Square Error (RMSE) and Mean Absolute Percentage Error (MAPE) from the predicted values of the testing data.}
\usage{
doKNN(train, test, knum)
}
\arguments{
\item{train}{Dataframe of training data, the first column is the response factor to be modelled by the remaining columns}

\item{test}{Dataframe of testing data, the first column is the response factor, the rest of the columns contain factors used to predict the response column with the built model}

\item{knum}{The number of K-Centroids used in the K-Nearest Neighbours method (see \code{\link[FNN]{knn.reg}})}
}
\value{
Returns:
\itemize{
\item mape - the Mean Absolute Percentage Error (MAPE) of the current model
\item rmse - the Root Mean Square Error (RMSE) of the current model
\item bestK - the value of K which returned the best RMSE and MAPE values
}
}
\description{
doKNN Applies the K-Nearest Neighbours Regression model to the training data,
and calculates the Root Mean Square Error (RMSE) and Mean Absolute Percentage Error (MAPE) from the predicted values of the testing data.
}
\details{
This function performs K-Nearest Neighbours regression \code{\link[FNN]{knn.reg}} on the training data, for each K specified.
The testing data is used to extract a Root Mean Square Error (RMSE) and Mean Absolute Percentage Error (MAPE) with the \code{\link[StompR]{getModelError}} function.
Only the lowest RMSE and corresponding MAPE and best K value are returned.
}
