% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRLR.R
\name{doRLR}
\alias{doRLR}
\title{doRLR Applies a Robust Linear Regression (RLR) model to the training data,
and calculates the Root Mean Square Error (RMSE) and Mean Absolute Percentage Error (MAPE) from the predicted values of the testing data.}
\usage{
doRLR(train, test, mscale)
}
\arguments{
\item{train}{Dataframe of training data, the first column is the response factor to be modelled by the remaining columns}

\item{test}{Dataframe of testing data, the first column is the response factor, the rest of the columns contain factors used to predict the response column with the built model}

\item{mscale}{The maximum number of iterations for the Iteratively Reweighted Least Squares algorithm; default is 50}
}
\value{
Returns the output of the function getModelError, which is a list consisting of:
\itemize{
\item mape - the Mean Absolute Percentage Error (MAPE) of the current model
\item rmse - the Root Mean Square Error (RMSE) of the current model
}
}
\description{
doRLR Applies a Robust Linear Regression (RLR) model to the training data,
and calculates the Root Mean Square Error (RMSE) and Mean Absolute Percentage Error (MAPE) from the predicted values of the testing data.
}
\details{
This function applies the RLR algorithm to the training data (see \code{\link[robustbase]{lmrob}}),
then calculates the RMSE and MAPE with \code{\link[StompR]{getModelError}} using the predicted values calculated from the testing data.
}
