% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSVM.R
\name{doSVM}
\alias{doSVM}
\title{doSVM Applies the Support Vector Machine algorithm (see \code{\link[e1071]{svm}}),
and then calculates the RMSE and MAPE with \code{\link[StompR]{getModelError}} using the predicted values calculated from the testing data.}
\usage{
doSVM(train, test, gamma, epsilon, cost)
}
\arguments{
\item{train}{Dataframe of training data, the first column is the response factor to be modelled by the remaining columns}

\item{test}{Dataframe of testing data, the first column is the response factor, the rest of the columns contain factors used to predict the response column with the built model}

\item{gamma}{Parameter of the Gaussian Kernel for non-linear classification; default is 1/(data dimension). This can be a list of values}

\item{epsilon}{The margin of tolerance for the epsilon-insensitive loss function; default is 0.1. This can be a list of values}

\item{cost}{The cost of misclassification; default is 1. This can be a list of values}
}
\value{
Returns the output of the function getModelError, which is a list consisting of:
\itemize{
\item mape - the Mean Absolute Percentage Error (MAPE) of the best model
\item rmse - the Root Mean Square Error (RMSE) of the best model
\item gamma - the gamma used for the best model
\item epsilon - the epsilon used for the best model
\item cost - the cost used for the best model
}
}
\description{
doSVM Applies the Support Vector Machine algorithm (see \code{\link[e1071]{svm}}),
and then calculates the RMSE and MAPE with \code{\link[StompR]{getModelError}} using the predicted values calculated from the testing data.
}
\details{
This function iterates over each cost, epsilon, and gamma value given, applying the SVM algorithm to the training data (see \code{\link[e1071]{svm}}).
Calculating the RMSE and MAPE with \code{\link[StompR]{getModelError}} using the predicted values calculated from the testing data,
the RMSE, MAPE, cost, gamma, and epsilon of the best performing model is returned.
}
