% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFactors.R
\name{extractFactors}
\alias{extractFactors}
\title{extractFactors creates a correlogram of the xMatrix data, and a scree plot, then returns the factors with the highest eigenvalues}
\usage{
extractFactors(file_path, xMatrix, group, title, cent, sca)
}
\arguments{
\item{file_path}{file path to the folder where the images are stored}

\item{xMatrix}{matrix of data}

\item{title}{name of the matrix to append to title}
}
\description{
extractFactors creates a correlogram of the xMatrix data, and a scree plot, then returns the factors with the highest eigenvalues
}
\details{
Takes a matrix of data for which each column is a continous factor and extracts the factors for which the eigenvalues are higher than 1. Also generates a scree plot and a correlogram.
}
