% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelScale.R
\name{levelScale}
\alias{levelScale}
\title{levelScale uses the Level Normalisation method to scale a matrix}
\usage{
levelScale(xMatrix)
}
\arguments{
\item{xMatrix}{matrix of data to be scaled}
}
\value{
returns the variable scalex, which is the scaled matrix
}
\description{
levelScale uses the Level Normalisation method to scale a matrix
}
\details{
uses the equation x' = (x - mean(x))/mean(x) to level scale the input matrix.
}
