% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meannormScale.R
\name{meannormScale}
\alias{meannormScale}
\title{meannormScale uses the Mean Normalisation method to scale a matrix of data}
\usage{
meannormScale(xMatrix)
}
\arguments{
\item{xMatrix}{matrix of data to be scaled}
}
\value{
returns the variable scalex, which is the scaled matrix
}
\description{
meannormScale uses the Mean Normalisation method to scale a matrix of data
}
\details{
This function uses the equation x' = (x - mean(x))/(max(x) - min(x)) to scale the matrix.
}
