% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmaxScale.R
\name{minmaxScale}
\alias{minmaxScale}
\title{minmaxScal uses the Min-Max Normalisation method to scale the input matrix}
\usage{
minmaxScale(xMatrix)
}
\arguments{
\item{xMatrix}{matrix of data to be scaled}
}
\value{
returns the variable scalex, which is the scaled matrix
}
\description{
minmaxScal uses the Min-Max Normalisation method to scale the input matrix
}
\details{
This function uses the equation x' = (x - min(x))/(max(x) - min(x)) to scale the matrix.
}
