% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoScale.R
\name{paretoScale}
\alias{paretoScale}
\title{paretoScale uses the Pareto Normalisation method to scale the input matrix}
\usage{
paretoScale(xMatrix)
}
\arguments{
\item{xMatrix}{matrix of data to be scaled}
}
\value{
returns the variable scalex, which is the scaled matrix
}
\description{
paretoScale uses the Pareto Normalisation method to scale the input matrix
}
\details{
This function uses the equation x' = (x - mean(x))/sqrt(sd(x)) to scale the input matrix.
}
