% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData}
\alias{splitData}
\title{splitData Splits the Data Sets into Training and Testing data}
\usage{
splitData(xMatrix, yVector, proportion)
}
\arguments{
\item{xMatrix}{a matrix which assumes the columns are factors, and the rows are samples}

\item{yVector}{a vector which should be the same length as}

\item{proportion}{the proportion of rows to keep in the training dataset to build the model, value between 0 and 1}
}
\value{
returns a list of the following:
\itemize{
\item train, a dataframe consisting of:
\itemize{
\item train.x - a matrix with rows from the xMatrix input to build the model
\item train.y - a vector with the corresponding yVector rows to train.x
}
\item test, a dataframe consisting of:
\itemize{
\item test.x - a matrix with rows from the xMatrix input to test the model performance
\item test.y - a vector with the corresponding yVector rows to test.x
}
}
}
\description{
splitData Splits the Data Sets into Training and Testing data
}
\details{
Uses the proportion to generate an index vector of random integers between 1 and proportion*length(yVector).
Then xMatrix is split using the index vector to extract rows, and stored as train.x and test.x.
The same is done to the yVector to generate train.y and test.y.
These are then combined into dataframs, where train contains train.x and train.y, and test contains test.x and test.y.
These two dataframes are returned in a list.
}
