% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startStomping.R
\name{startStomping}
\alias{startStomping}
\title{startStomping is a function which statistically compares a variety of models on the same datasets}
\usage{
startStomping(
  file_path,
  xMatrix,
  yVector,
  logV,
  transformV,
  meth,
  prop,
  seed,
  iter,
  plsr_ncomp,
  rfr_ntree,
  svm_gamma,
  svm_epsilon,
  svm_cost,
  knn_knum,
  gbm_ntree,
  gbm_shrink,
  gbm_dist,
  gbm_node,
  rlr_mscale,
  permission
)
}
\arguments{
\item{file_path}{string String to a folder where the output plots are stored}

\item{xMatrix}{A matrix where each column is considered a factor to be modelled. Names of columns will automatically be used if provided.}

\item{yVector}{A vector containing the response factor to be modelled}

\item{logV}{Boolean value, if TRUE then transform the xMatrix using log to base 10}

\item{transformV}{Transformation and scaling method to be applied to xMatrix, raw data is used by default, the options are: raw, center, minmax, meannorm, zscore, pareto, vast, level}

\item{meth}{Numerical vector of statistical methods to be compared, all of them are used by default if left blank. c(1:11) are indices of methods: "OLSR", "SLR-both", "SLR-forward", "SLR-backward", "PCR", "PLSR", "RFR", "SVM", "KNN", "GBM", "GLMR"}

\item{prop}{Proportion of data to be used in the training of the model. Value between 0 and 1. For example, 0.7 is 70\% of the data used for training the model}

\item{seed}{Initial seed for splitting training and testing datasets; used for reproducability of results}

\item{iter}{How many models are built to assess the overall accuracy of the method}

\item{plsr_ncomp}{The number of components used for the principal components regression method}

\item{rfr_ntree}{The number of trees used in the Random Forest Regression method}

\item{svm_gamma}{The gamma used for the Support Vector Machine method}

\item{svm_epsilon}{The epsilon used for the Support Vector Machine method}

\item{svm_cost}{The cost used for the Support Vector Machine Method}

\item{knn_knum}{The number of K-Centroids used in the K-Nearest Neighbours method}

\item{gbm_ntree}{The number of trees used in the Generalised Boosted regression method}

\item{gbm_shrink}{The shrink used in the Generalised Boosted regression method}

\item{gbm_dist}{The distribution used in the Generalised Boosted regression method}

\item{gbm_node}{The number of nodes used in the Generalised Boosted regression method}

\item{rlr_mscale}{Integer specifying the maximum number of iterations for the IRWLS algorithm within the Robust Linear regression method; default is 50}

\item{permission}{true,false Permission for this package to create files in the file path specified}
}
\value{
The performance plots are in the file path specified. Also returned is a list of the following:
\itemize{
\item RMSE_CM - The cumulative mean values for the RMSE, the final value is the converged mean of the RMSE of the model
\item MAPE_CM - The cumulative mean values for the MAPE, the final value is the converged mean of the MAPE of the model
\item rmse_raw - The raw RMSE value of each model for each iteration
\item mape_raw - The raw MAPE value of each model for each iteration
}
}
\description{
startStomping is a function which statistically compares a variety of models on the same datasets
}
\details{
This function returns the error rates of each model vs the data supplied.
There are two measurements of error, the root mean square error (RMSE), and the mean absolute error (MAPE).
The models that this function compares are:
\itemize{
\item Ordinary Least Squares Regression (OLSR)
\item Stepwise Linear Regression - Both directions (SLR)
\item Stepwise Linear Regression - Forwards direction (SLRf)
\item Stepwise Linear Regression - Backwards direction (SLRb)
\item Principal Components Regression (PCR)
\item Partial Least Squares Regression (PLSR)
\item Random Forest Regression (RFR)
\item Support Vector Machine (SVM)
\item K-Nearest Neighbours Regression (KNN)
\item Generalised Boosted Modelling (GBM)
\item Robust Linear Regression (RLR)
}
}
