% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startStompingMultipleY.R
\name{startStompingMultipleY}
\alias{startStompingMultipleY}
\title{startStompingMultipleY This function compares different statistical methods as well as different xMatrices vs the different yVectors response factors.}
\usage{
startStompingMultipleY(
  file_path,
  xMatrices,
  yVectors,
  logV,
  transformV,
  meth,
  prop,
  seed,
  iter,
  plsr_ncomp,
  rfr_ntree,
  svm_gamma,
  svm_epsilon,
  svm_cost,
  knn_knum,
  gbm_ntree,
  gbm_shrink,
  gbm_dist,
  gbm_node,
  rlr_mscale,
  permission = F
)
}
\arguments{
\item{file_path}{string String to a folder where the output plots are stored.}

\item{xMatrices}{A list of matrices where each column is considered a factor to be modelled. Names of columns will automatically be used if provided, and the names in the list will be used as the name of that dataset.}

\item{yVectors}{A list of vectors containing the response factors to be modelled}

\item{logV}{Boolean value, if TRUE then transform the xMatrix using log to base 10}

\item{transformV}{Transformation and scaling method to be applied to xMatrix, raw data is used by default, the options are: raw, center, minmax, meannorm, zscore, pareto, vast, level}

\item{meth}{Numerical vector of statistical methods to be compared, all of them are used by default if left blank. c(1:11) are indices of methods: "OLSR", "SLR-both", "SLR-forward", "SLR-backward", "PCR", "PLSR", "RFR", "SVM", "KNN", "GBM", "GLMR"}

\item{prop}{Proportion of data to be used in the training of the model. Value between 0 and 1. For example, 0.7 is 70\% of the data used for training the model.}

\item{seed}{Initial seed for splitting training and testing datasets. Used for reproducability of results.}

\item{iter}{How many models are built to assess the overall accuracy of the method}

\item{plsr_ncomp}{The number of components used for the principal components regression method}

\item{rfr_ntree}{The number of trees used in the Random Forest Regression method}

\item{svm_gamma}{The gamma used for the Support Vector Machine method}

\item{svm_epsilon}{The epsilon used for the Support Vector Machine method}

\item{svm_cost}{The cost used for the Support Vector Machine Method}

\item{knn_knum}{The number of K-Centroids used in the K-Nearest Neighbours method}

\item{gbm_ntree}{The number of trees used in the Generalised Boosted regression method}

\item{gbm_shrink}{The shrink used in the Generalised Boosted regression method}

\item{gbm_dist}{The distribution used in the Generalised Boosted regression method}

\item{gbm_node}{The number of nodes used in the Generalised Boosted regression method}

\item{rlr_mscale}{The mscale used for the Robust Linear regression method}

\item{permission}{true,false Permission for this package to create files in the file path specified.}
}
\value{
nothing yet
}
\description{
startStompingMultipleY This function compares different statistical methods as well as different xMatrices vs the different yVectors response factors.
}
\details{
A time-saving method of analysing different xMatrix datasets against different yVector response factors.
It also compares each xMatrix versus a variety of different statistical models; the available models to compare are:
\itemize{
\item Ordinary Least Squares Regression (OLSR)
\item Stepwise Linear Regression - Both directions (SLR)
\item Stepwise Linear Regression - Forwards direction (SLRf)
\item Stepwise Linear Regression - Backwards direction (SLRb)
\item Principal Components Regression (PCR)
\item Partial Least Squares Regression (PLSR)
\item Random Forest Regression (RFR)
\item Support Vector Machine (SVM)
\item K-Nearest Neighbours Regression (KNN)
\item Generalised Boosted Modelling (GBM)
\item Robust Linear Regression (RLR)
}
There are two methods of analysing the performance of each model, these are the Root Mean Square Error (RMSE) and the Mean Absolute Percentage Error (MAPE);
there are two heatmaps produced and stored as files in the file path specified, one for each of the perfomance analysis methods.
If permission is not specified as true, it will request a user input to confirm consent to create images on the user computer, this step can be skipped/automated by setting that parameter to true.
}
