% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vastScale.R
\name{vastScale}
\alias{vastScale}
\title{vastScale uses the Variable Stability Scaling Transformation (VAST) to normalise the input matrix}
\usage{
vastScale(xMatrix)
}
\arguments{
\item{xMatrix}{matrix of data to be scaled}
}
\value{
returns the variable scalex, which is the scaled matrix
}
\description{
vastScale uses the Variable Stability Scaling Transformation (VAST) to normalise the input matrix
}
\details{
This function uses the equation x' = ((x - mean(x)) / sd(x)) . (mean(x)/sd(x)) to normalise the input matrix.
}
