% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreScale.R
\name{zscoreScale}
\alias{zscoreScale}
\title{zscoreScale uses the Z-Score Normalisation method to scale the input matrix}
\usage{
zscoreScale(xMatrix)
}
\arguments{
\item{xMatrix}{matrix of data to be scaled}
}
\value{
returns the variable scalex, which is the scaled matrix
}
\description{
zscoreScale uses the Z-Score Normalisation method to scale the input matrix
}
\details{
This function uses the equation x' = (x - mean(x))/sd(x) to normalise the input matrix
}
