% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelError.R
\name{getModelError}
\alias{getModelError}
\title{getModelError is a function that calculates the Root Mean Square Error (RMSE) and Mean Absolute Percentage Error (MAPE) from the Predicted values and corresponding Real values}
\usage{
getModelError(Predicted, Actual)
}
\arguments{
\item{Predicted}{The output of the function \code{\link[stats]{predict}}}

\item{Actual}{The corresponding real values to the testing data used to calculate the Predicted values}
}
\value{
Returns:
\itemize{
\item mape - the Mean Absolute Percentage Error (MAPE) of the current model
\item rmse - the Root Mean Square Error (RMSE) of the current model
}
If they are infinite values, this function changes them to 999 to avoid Inf errors
}
\description{
getModelError is a function that calculates the Root Mean Square Error (RMSE) and Mean Absolute Percentage Error (MAPE) from the Predicted values and corresponding Real values
}
